+++ src/controllers/inventory/product/deleteProduct.ts
@@ -8,22 +8,26 @@ export const deleteProduct = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
 
+    const relatedDetails = await prisma.productDetail.findMany({
+      where: { productId:id },
+    });
+    
+    if (relatedDetails.length > 0) {
+      return res.status(400).json({
+        message: 'Cannot delete product. Related details associated with invoice.',
+      });
+    }
+    
+
     await prisma.product.delete({
       where: { id },
     });
 
     return res.status(200).json({ message: "Product deleted successfully" });
   } catch (error:any) {
-    if (
-      error.code === "P2003" // Prisma error code for foreign key constraint
-    ) {
-      res.status(400).json({
-        error:
-          "Cannot delete category as it is associated with one or more products.",
-      });
-    } else {
+    
     res.status(500).json({ message: "Error deleting product", error });
     return error;
-  }
+  
 }
 };
